+++ packages/headless/src/features/sort-criteria/sort-criteria-insight-analytics-actions.ts
@@ -2,6 +2,7 @@ import {
   AnalyticsType,
   makeInsightAnalyticsAction,
 } from '../analytics/analytics-utils';
+import {getCaseContextAnalyticsMetadata} from '../case-context/case-context-state';
 import {getSortCriteriaInitialState} from './sort-criteria-state';
 
 export const logResultsSort = makeInsightAnalyticsAction(
@@ -10,5 +11,6 @@ export const logResultsSort = makeInsightAnalyticsAction(
   (client, state) =>
     client.logResultsSort({
       resultsSortBy: state.sortCriteria || getSortCriteriaInitialState(),
+      ...getCaseContextAnalyticsMetadata(state.insightCaseContext),
     })
 );