+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.9](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.8...@coveo/atomic-angular@1.30.9) (2022-09-20)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.8](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.7...@coveo/atomic-angular@1.30.8) (2022-09-20)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.7](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.6...@coveo/atomic-angular@1.30.7) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.6](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.5...@coveo/atomic-angular@1.30.6) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.5](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.4...@coveo/atomic-angular@1.30.5) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.3...@coveo/atomic-angular@1.30.4) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.2...@coveo/atomic-angular@1.30.3) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [1.30.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.1...@coveo/atomic-angular@1.30.2) (2022-09-16)
 
 **Note:** Version bump only for package @coveo/atomic-angular