+++ packages/atomic/CHANGELOG.md
@@ -3,6 +3,82 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.113.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.113.0...@coveo/atomic@1.113.1) (2022-09-20)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.113.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.4...@coveo/atomic@1.113.0) (2022-09-20)
+
+
+### Features
+
+* **atomic:** initial work for recommendations list and interface ([#2454](https://github.com/coveo/ui-kit/issues/2454)) ([7ba8de1](https://github.com/coveo/ui-kit/commit/7ba8de143b9bc281449c1820480695ebe5375764))
+
+
+
+
+
+## [1.112.4](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.3...@coveo/atomic@1.112.4) (2022-09-19)
+
+
+### Bug Fixes
+
+* **atomic:** allow facets to be re-added to the interface ([#2450](https://github.com/coveo/ui-kit/issues/2450)) ([145cec4](https://github.com/coveo/ui-kit/commit/145cec4df503abc417076af5d4d55070ef105324))
+* **atomic:** prevent scrolling behind refine modal on iOS ([#2451](https://github.com/coveo/ui-kit/issues/2451)) ([f270912](https://github.com/coveo/ui-kit/commit/f270912d65ad28fc8788a0cbba5ac8f7632e62a1))
+* **atomic:** removed wrong section class from results ([#2452](https://github.com/coveo/ui-kit/issues/2452)) ([9011f57](https://github.com/coveo/ui-kit/commit/9011f57a5b0556f0438f00f818c66f3b2bdd5445))
+* **atomic:** use correct delimiting character for category facet headless controller ([#2439](https://github.com/coveo/ui-kit/issues/2439)) ([246e140](https://github.com/coveo/ui-kit/commit/246e140c97e84c7943e8d92ef89429fea4c5e817))
+
+
+
+
+
+## [1.112.3](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.2...@coveo/atomic@1.112.3) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+## [1.112.2](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.1...@coveo/atomic@1.112.2) (2022-09-19)
+
+
+### Bug Fixes
+
+* **atomic:** fix default template for atomic-result-text on language ([#2447](https://github.com/coveo/ui-kit/issues/2447)) ([6b24146](https://github.com/coveo/ui-kit/commit/6b2414644a871e0e39bf0791b3a7f51908241344))
+
+
+
+
+
+## [1.112.1](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.112.0...@coveo/atomic@1.112.1) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/atomic
+
+
+
+
+
+# [1.112.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.111.0...@coveo/atomic@1.112.0) (2022-09-16)
+
+
+### Bug Fixes
+
+* **atomic:** render result list content when it's ready ([#2438](https://github.com/coveo/ui-kit/issues/2438)) ([0f62dcd](https://github.com/coveo/ui-kit/commit/0f62dcd4ea66b35033fc1199b59f216cbe718b02))
+
+
+### Features
+
+* **atomic:** added support for multiple notifications ([#2437](https://github.com/coveo/ui-kit/issues/2437)) ([a40dce2](https://github.com/coveo/ui-kit/commit/a40dce2cfa68e101e0e03d901673f192468c1993))
+
+
+
+
+
 # [1.111.0](https://github.com/coveo/ui-kit/compare/@coveo/atomic@1.110.1...@coveo/atomic@1.111.0) (2022-09-16)