+++ packages/quantic/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.44.8](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.7...@coveo/quantic@1.44.8) (2022-09-20)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.44.7](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.6...@coveo/quantic@1.44.7) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.44.6](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.5...@coveo/quantic@1.44.6) (2022-09-19)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.44.5](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.4...@coveo/quantic@1.44.5) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
+## [1.44.4](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.3...@coveo/quantic@1.44.4) (2022-09-16)
+
+**Note:** Version bump only for package @coveo/quantic
+
+
+
+
+
 ## [1.44.3](https://github.com/coveo/ui-kit/compare/@coveo/quantic@1.44.2...@coveo/quantic@1.44.3) (2022-09-16)
 
 **Note:** Version bump only for package @coveo/quantic